+++ app/Http/Controllers/ItemController.php
@@ -126,6 +126,28 @@ class ItemController extends Controller
         $form->handleRequest();
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $authors = $form['author']->getData();
+
+            
+            $unknowns = array_filter($authors, function ($a) {
+                return !is_numeric($a);
+            });
+            if (sizeof($authors) > 0) {
+                $orders = array_map(function ($item, $key) {
+                    return ['order' => $key];
+                }, $authors, range(1, sizeof($authors)));
+
+                $item->authorities()->sync(array_combine($authors, $orders));
+            } else {
+                $item->authorities()->sync([]);
+            }
+            $authorities = $item->authorities()->orderBy('order', 'asc')->pluck('name', 'id');
+            $item->author =  implode(
+                ';',
+                array_map(function($a) use ($authorities){
+                    return isset($authorities[$a])?$authorities[$a] : $a;
+                }, $authors)
+            );
             $item->push();
 
             $tags = $form['tags']->getData();